\name{geneNames}
\alias{geneNames}
\docType{data}
\title{
Gene Names
}
\description{
A vector of gene names. The size of the vector should be equal to the number of rows in your data sets. 
}
\usage{data("geneNames")}
\format{
  A data frame with 0 observations on the following 2 variables.
  \describe{
    \item{\code{x}}{a character vector}
    \item{\code{y}}{a numeric vector}
  }
}

\source{
http://www.broadinstitute.org/cgi-bin/cancer/datasets.cgi
}
\references{
Singh,D., Febbo,P.G., Ross,K., et al. (2002) Gene expression correlates of clinical prostatecancer behavior. Cancer cell. 1:203-9.
}
\examples{

data(geneNames)

}
\keyword{datasets}
