\name{normalData}
\alias{normalData}
\docType{data}
\title{
Example Prostate Cancer Normal Gene Expression Dataset
}
\description{
Raw microarray gene expression data set. This data is randomly selected 500 gene subset of the original prostate cancer data set taken from Broad Institute.
}
\usage{data("normalData")}

\details{
The example data setnis also given to provide a sample data set so that one can take it as reference while formatting the new data set as entry to dc3net. Data sets must be complete and no empty or non-numerical character allowed. 
}
\source{
This data was taken from http://www.broadinstitute.org/cgi-bin/cancer/datasets.cgi
}
\references{
Singh,D., Febbo,P.G., Ross,K., et al. (2002) Gene expression correlates of clinical prostatecancer behavior. Cancer cell. 1:203-9.
}
\examples{

data(normalData)

}
\keyword{datasets}
