\name{orderList}
\alias{orderList}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
  Order The List of Interactions
}
\description{
\code{orderList} takes the data matrix generated by maxofMIM function and sort it according to mutual information values in increasing or decreasing order.
}
\usage{
orderList(mylist, colindx=1, decreasing=TRUE)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{mylist}{The data matrix generated by maxofMIM function that rows are probe/gene names and columns are mutual information values, row numbers of each gene and row numbers of partner gene with highest mutual information value, respectively. }
  \item{colindx}{The number of column with mutual information values}
  \item{decreasing}{The order of sort. If the argument is set to FALSE, then the data matrix will be sort in increasing order.}
}
\details{
\code{orderList} takes the data matrix generated by maxofMIM function and sort it according to the column number stated in colindx argument. The direction of order is set by decreasing argument.
}
\value{
  \code{orderList} returns a sorted data matrix that rows are probe/gene names and columns are mutual information values, row numbers of each gene and row numbers of partner gene with highest mutual information value, respectively.
}
\references{
 Altay, G. et. al. "Genome-wide differential gene network analysis software and an application  in LnCap prostate cancer", bioRxiv, 2017. 
  Altay, G., Asim, M., Markowetz, F., Neal, D.E. (2011) Differential C3NET reveals disease netwokrs of direct physical interactions. BMC Bioinformatics. 12: 296.
}
\author{
  Gokmen Altay
}

\seealso{
\code{\link{dc3net}}, \code{\link{maxofMIM}},
}

\keyword{misc}
